refactor(frontend): simplify folder structure by removing the id folder from employee folder (#288)
remove the id folder from employee folder to simplify the folder structure
