+++ frontend/app/i18n-routes.ts
@@ -130,7 +130,7 @@ export const i18nRoutes = [
       },
       {
         id: 'EMPL-0003',
-        file: 'routes/employee/[id]/profile/index.tsx',
+        file: 'routes/employee/profile/index.tsx',
         paths: {
           en: '/en/employee/:id/profile',
           fr: '/fr/employe/:id/profil',
@@ -138,7 +138,7 @@ export const i18nRoutes = [
       },
       {
         id: 'EMPL-0004',
-        file: 'routes/employee/[id]/profile/personal-information.tsx',
+        file: 'routes/employee/profile/personal-information.tsx',
         paths: {
           en: '/en/employee/:id/profile/personal-information',
           fr: '/fr/employe/:id/profil/informations-personnelles',
@@ -146,7 +146,7 @@ export const i18nRoutes = [
       },
       {
         id: 'EMPL-0005',
-        file: 'routes/employee/[id]/profile/employment-information.tsx',
+        file: 'routes/employee/profile/employment-information.tsx',
         paths: {
           en: '/en/employee/:id/profile/employment-information',
           fr: `/fr/employe/:id/profil/informations-sur-lemploi`,
@@ -154,7 +154,7 @@ export const i18nRoutes = [
       },
       {
         id: 'EMPL-0006',
-        file: 'routes/employee/[id]/profile/referral-preferences.tsx',
+        file: 'routes/employee/profile/referral-preferences.tsx',
         paths: {
           en: '/en/employee/:id/profile/referral-preferences',
           fr: '/fr/employe/:id/profil/préférences-de-référence',
