+++ frontend/app/routes/employee/profile/personal-information.tsx
@@ -4,7 +4,7 @@ import { data, Form } from 'react-router';
 import { Trans, useTranslation } from 'react-i18next';
 import * as v from 'valibot';
 
-import type { Route } from './+types/personal-information';
+import type { Route } from '../profile/+types/personal-information';
 
 import type { Profile } from '~/.server/domain/models';
 import { getLanguageForCorrespondenceService } from '~/.server/domain/services/language-for-correspondence-service';
@@ -22,7 +22,7 @@ import { InputTextarea } from '~/components/input-textarea';
 import { InlineLink } from '~/components/links';
 import { HttpStatusCodes } from '~/errors/http-status-codes';
 import { getTranslation } from '~/i18n-config.server';
-import { personalInformationSchema } from '~/routes/employee/[id]/profile/validation.server';
+import { personalInformationSchema } from '~/routes/employee/profile/validation.server';
 import { handle as parentHandle } from '~/routes/layout';
 import { toE164 } from '~/utils/phone-utils';
 import { formString } from '~/utils/string-utils';
@@ -66,7 +66,7 @@ export async function action({ context, params, request }: Route.ActionArgs) {
     throw updateResult.unwrapErr();
   }
 
-  return i18nRedirect('routes/employee/[id]/profile/index.tsx', request, {
+  return i18nRedirect('routes/employee/profile/index.tsx', request, {
     params: { id: currentUserId },
   });
 }
@@ -109,7 +109,7 @@ export default function PersonalInformation({ loaderData, actionData, params }: 
 
   return (
     <>
-      <InlineLink className="mt-6 block" file="routes/employee/[id]/profile/index.tsx" params={params} id="back-button">
+      <InlineLink className="mt-6 block" file="routes/employee/profile/index.tsx" params={params} id="back-button">
         {`< ${t('app:profile.back')}`}
       </InlineLink>
       <div className="max-w-prose">
@@ -204,12 +204,7 @@ export default function PersonalInformation({ loaderData, actionData, params }: 
                 <Button name="action" variant="primary" id="save-button">
                   {t('app:form.save')}
                 </Button>
-                <ButtonLink
-                  file="routes/employee/[id]/profile/index.tsx"
-                  params={params}
-                  id="cancel-button"
-                  variant="alternative"
-                >
+                <ButtonLink file="routes/employee/profile/index.tsx" params={params} id="cancel-button" variant="alternative">
                   {t('app:form.cancel')}
                 </ButtonLink>
               </div>
